backstop: chore: remove unnecessary code seperation in backstop bad debt flow
