+++ pool/src/pool/mod.rs
@@ -2,7 +2,7 @@ mod actions;
 pub use actions::Request;
 
 mod bad_debt;
-pub use bad_debt::{burn_backstop_bad_debt, transfer_bad_debt_to_backstop};
+pub use bad_debt::transfer_bad_debt_to_backstop;
 
 mod config;
 pub use config::{
